Two-pass lookup mechanism for i18n (#351).
The first pass does a lookup on the original (english) message and tries to
find a msgid for it (which is some sort of internal code for the message,
eg. "memory.used"). This msgid is then used in the second lookup, this time
against the real translation catalog (eg. to retrieve "Memory used:").

This mechanism, while a little more demanding on resources, makes it possible
to abstract ourselves from the actual message used in the code.
Whenever the message in the code changes (eg. a typo gets fixed), the
intermediate catalog can be updated to reflect the change.
The final catalogs don't need to be updated because the msgids need not be
changed.

I'm making this change as a little experiment, to find a good balance between
readability, code abstraction & effectiveness. Most of gettext's tools can
still be used with this mechanism and the other tools can be written as very
simple shell scripts, eg.:
grep msgid .../module.po | sed -e 's/msgid[^"]+"\(.*\)"/\1/'