Ability to distinguish between connections using the network's name in the console (#78).
When sending a command to the bot so it is relayed to some IRC server, you must now prefix the message
with a pattern matching the name of the network you want the message sent to, followed by a single space.
Wildcards ('*' and '?') are supported.
Eg. to send a message to DALnet and supposing a "DALnet" network was configured,
I could pass the following command to the console's socket (without quotes):
"DALnet PRIVMSG #test :Some message\r\n"
or just:
"* PRIVMSG #test :Some message\r\n"
which will send the same message to all the servers the bot is currently connected to.