Make Erebot_Config_Main more robust.
The class now accepts instances of Erebot_URI as its $configData.
It also does a better job at checking the validity of its arguments.
Last but not least, if $source is LOAD_FROM_FILE and $configData points
to a relative path, it is now interpreted as relative to the current
working directory instead of some PEAR-based directory. The former
behaviour did not make much sense anyway.