Fix issues with timers and xglob streams.
The command used to start timers on Windows used improper quoting and relied
on the wrong tools (start.exe & php-win.exe).

Callback names are now filtered through addslashes() before being added to
the timer's command line. This should prevent a possible vulnerability when
namespaces are used (as the backslash character may be misinterpreted by the
shell).

The xglob user-defined stream failed to create absolute paths under Windows
when the input started with a backslash or a drive letter (eg. "C:/*.xml").
As of now, forward slashes (/) are treated as directory separator in
addition to the operating system's regular DIRECTORY_SEPARATOR.
This change makes the xglob stream more in line with other URI-based schemes.

All-in-all, this should make an additional 2 tests pass on Windows.