Fix more little issues.
Cosmetic touch: make Erebot_DOM create LibXMLErrors with members appearing
in the same order as when it's raised internally by PHP.

Don't hard-code line-endings in i18n msgids to ensure maximum portability
(especially with regard to Windows).

Work around a strange bug with timers under Windows where a process is
marked as read-ready right after is was created (with popen()).
Thus, it is necessary to read() from the stream representing the process
as soon as possible to reset the read-ready status.

Don't display warnings from stream_select() in IdentD tests. This should
work around the following duplicate bugs:
* https://bugs.php.net/bug.php?id=49948
* https://bugs.php.net/bug.php?id=54563
... and hopefully make these tests pass on Windows too.

Don't mask errors completely before using File_Gettext, especially not
fatal errors. This makes it easier to debug certain bugs, eg. when
File_Gettext is missing from the system's pear installation.
Also, don't assign the result of File_Gettext::factory() by reference
as this is deprecated now (but a cleanup will probably be needed in
File_Gettext itself to get rid of other similar issues and remove the
need for error_reporting() changes).

Refactor the TimerTest a little to make it easier to read and maintain.