Better handling for timers on Windows.
Most of the timers' logic has been rewritten to make it work on Windows.
The simpler approach previously used (based on popen()) doesn't seem to
work for Windows.

We now use two different strategies depending on the OS:
* On Windows, a temporary file is created and the subprocess writes to
  it when it's done (when the timer triggers). The temporary file's
  file descriptor is used by select().
* On other platforms, the subprocess' stdout is replaced by a pipe,
  which is marked as read-ready when the subprocess exits (after the
  timer triggers). That pipe's file descriptor is used by select().

Also, the class now does a better job at finding the PHP binary to call
(it caches the value so that subsequent timers or invocations of the
same timer don't have to look for this information again).