A few renaming.
Console_CommandLine_MyOption -> Console_CommandLine_ParallelOption because
it makes it possible for 2 concurrent options to work in parallel on the
same variable.

Erebot_Console -> Erebot_Prompt because the previous name was a little
counter-intuitive : the so-called "Console" can only be used to write
stuff for Erebot to handle, it cannot be used for read access.

Other adjustments were needed for these changes to work.