Exclude the styles' parser from doxygen.
That Parser.php is auto-generated and its API is documented using a different
tool, triggering a lot of warnings when parsed by doxygen.
The doc in that file is of no interest to us anyway, as developpers interested
in changing the parser should do so through its grammar (see data/Styling.y)
instead.