Fix warnings about risky tests
PHPUnit emits a warning by default when a test does no assertions.

In our case, we test whether configuration files/strings can be loaded
by just attempting to load them. The mere fact that no exception is
thrown means that the test is successul. The lack of further assertions
resulted in those warnings.

We now test a tautology in those test cases to silence PHPUnit.